Core/WorldStates: Migrate battleground worldstates to new system
Port From (https://github.com/TrinityCore/TrinityCore/commit/5dbfaecd5c020fd4c3b3dfc235580849c39ce7db)

+++ Source/Game/Entities/GameObject/GameObject.cs
@@ -3182,7 +3182,7 @@ namespace Game.Entities
                     packet.CapturePointInfo.CaptureTotalDuration = TimeSpan.FromMilliseconds(GetGoInfo().CapturePoint.CaptureTime);
                     packet.CapturePointInfo.CaptureTime = m_goValue.CapturePoint.AssaultTimer;
                     bg.SendPacketToAll(packet);
-                    bg.UpdateWorldState(GetGoInfo().CapturePoint.worldState1, (byte)m_goValue.CapturePoint.State);
+                    bg.UpdateWorldState((int)GetGoInfo().CapturePoint.worldState1, (byte)m_goValue.CapturePoint.State);
                 }
             }
         }
